refactored, added file_parser module
