+++ tests/outputs/output_bt.json
@@ -1,6 +1,125 @@
 {
   "name": "Level",
   "values": {
+    "intTest": {
+      "Int": {
+        "name": "intTest",
+        "value": 2147483647
+      }
+    },
+    "stringTest": {
+      "String": {
+        "name": "stringTest",
+        "value": "HELLO WORLD THIS IS A TEST STRING ÃÃÃ!"
+      }
+    },
+    "floatTest": {
+      "Float": {
+        "name": "floatTest",
+        "value": 0.49823147
+      }
+    },
+    "listTest (long)": {
+      "List": {
+        "name": "listTest (long)",
+        "ty": "Long",
+        "values": [
+          {
+            "Long": {
+              "name": "",
+              "value": 11
+            }
+          },
+          {
+            "Long": {
+              "name": "",
+              "value": 12
+            }
+          },
+          {
+            "Long": {
+              "name": "",
+              "value": 13
+            }
+          },
+          {
+            "Long": {
+              "name": "",
+              "value": 14
+            }
+          },
+          {
+            "Long": {
+              "name": "",
+              "value": 15
+            }
+          }
+        ]
+      }
+    },
+    "doubleTest": {
+      "Double": {
+        "name": "doubleTest",
+        "value": 0.4931287132182315
+      }
+    },
+    "longTest": {
+      "Long": {
+        "name": "longTest",
+        "value": 9223372036854775807
+      }
+    },
+    "shortTest": {
+      "Short": {
+        "name": "shortTest",
+        "value": 32767
+      }
+    },
+    "nested compound test": {
+      "Compound": {
+        "name": "nested compound test",
+        "values": {
+          "ham": {
+            "Compound": {
+              "name": "ham",
+              "values": {
+                "name": {
+                  "String": {
+                    "name": "name",
+                    "value": "Hampus"
+                  }
+                },
+                "value": {
+                  "Float": {
+                    "name": "value",
+                    "value": 0.75
+                  }
+                }
+              }
+            }
+          },
+          "egg": {
+            "Compound": {
+              "name": "egg",
+              "values": {
+                "name": {
+                  "String": {
+                    "name": "name",
+                    "value": "Eggbert"
+                  }
+                },
+                "value": {
+                  "Float": {
+                    "name": "value",
+                    "value": 0.5
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    },
     "listTest (compound)": {
       "List": {
         "name": "listTest (compound)",
@@ -10,16 +129,16 @@
             "Compound": {
               "name": "",
               "values": {
-                "created-on": {
-                  "Long": {
-                    "name": "created-on",
-                    "value": 1264099775885
-                  }
-                },
                 "name": {
                   "String": {
                     "name": "name",
                     "value": "Compound tag #0"
+                  }
+                },
+                "created-on": {
+                  "Long": {
+                    "name": "created-on",
+                    "value": 1264099775885
                   }
                 }
               }
@@ -51,51 +170,6 @@
       "Byte": {
         "name": "byteTest",
         "value": 127
-      }
-    },
-    "nested compound test": {
-      "Compound": {
-        "name": "nested compound test",
-        "values": {
-          "egg": {
-            "Compound": {
-              "name": "egg",
-              "values": {
-                "value": {
-                  "Float": {
-                    "name": "value",
-                    "value": 0.5
-                  }
-                },
-                "name": {
-                  "String": {
-                    "name": "name",
-                    "value": "Eggbert"
-                  }
-                }
-              }
-            }
-          },
-          "ham": {
-            "Compound": {
-              "name": "ham",
-              "values": {
-                "value": {
-                  "Float": {
-                    "name": "value",
-                    "value": 0.75
-                  }
-                },
-                "name": {
-                  "String": {
-                    "name": "name",
-                    "value": "Hampus"
-                  }
-                }
-              }
-            }
-          }
-        }
       }
     },
     "byteArrayTest (the first 1000 values of (n*n*255+n*7)%100, starting with n=0 (0, 62, 34, 16, 8, ...))": {
@@ -1103,80 +1177,6 @@
           6,
           48
         ]
-      }
-    },
-    "listTest (long)": {
-      "List": {
-        "name": "listTest (long)",
-        "ty": "Long",
-        "values": [
-          {
-            "Long": {
-              "name": "",
-              "value": 11
-            }
-          },
-          {
-            "Long": {
-              "name": "",
-              "value": 12
-            }
-          },
-          {
-            "Long": {
-              "name": "",
-              "value": 13
-            }
-          },
-          {
-            "Long": {
-              "name": "",
-              "value": 14
-            }
-          },
-          {
-            "Long": {
-              "name": "",
-              "value": 15
-            }
-          }
-        ]
-      }
-    },
-    "stringTest": {
-      "String": {
-        "name": "stringTest",
-        "value": "HELLO WORLD THIS IS A TEST STRING ÃÃÃ!"
-      }
-    },
-    "doubleTest": {
-      "Double": {
-        "name": "doubleTest",
-        "value": 0.4931287132182315
-      }
-    },
-    "shortTest": {
-      "Short": {
-        "name": "shortTest",
-        "value": 32767
-      }
-    },
-    "longTest": {
-      "Long": {
-        "name": "longTest",
-        "value": 9223372036854775807
-      }
-    },
-    "intTest": {
-      "Int": {
-        "name": "intTest",
-        "value": 2147483647
-      }
-    },
-    "floatTest": {
-      "Float": {
-        "name": "floatTest",
-        "value": 0.49823147
       }
     }
   }
